% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{abn2bnlearn.plot}
\alias{abn2bnlearn.plot}
\title{Plot ABN DAG with `bnlearn::graphviz.plot`}
\usage{
abn2bnlearn.plot(data, abndag)
}
\arguments{
\item{data}{Data frame used to learn abn DAG}

\item{abndag}{List from output of `abn::mostProbable()`}
}
\value{
plot
}
\description{
Plot ABN DAG with `bnlearn::graphviz.plot`
}
\examples{
\dontrun{
mycache = buildScoreCache(data.df = data, data.dists = dists, max.parents = npar, method = "mle",
dag.banned = banned, verbose = FALSE)
dag.mP = mostProbable(score.cache = mycache, score = "bic", verbose = FALSE)
abn2bnlearn.plot(data, dag.mP)
}
}
