% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{abnwithmaxparents}
\alias{abnwithmaxparents}
\title{Learn ABN with max parents.}
\usage{
abnwithmaxparents(
  net.scores,
  df,
  dist,
  banned,
  retain,
  filenamesuffix,
  filenamebase = FILENAMEbase,
  filename = FILENAME,
  method = METHOD,
  no.cores = n.cores,
  score = SCORE,
  randomeffect = NULL,
  catcov.restriction = "diagonal",
  rngSEED = SEED
)
}
\arguments{
\item{net.scores}{output from \code{bnaiaR::findOptNoParentNodes()$net.scores}}

\item{df}{data.frame with training data}

\item{dist}{list of variable = distribution type. Details in \code{?abn::buildScoreCache()}.}

\item{banned}{matrix with column variable as parent of row-variable, that defines which arcs are not permitted. Details in \code{?abn::buildScoreCache()}.}

\item{retain}{matrix with column variable as parent of row-variable, that defines which arcs are enforced. Details in \code{?abn::buildScoreCache()}.}

\item{filenamesuffix}{character specifying the current run (e.g. experiment number).}

\item{filenamebase}{character of path to location where results should be stored.}

\item{filename}{character specifying the current set of runs.}

\item{method}{either "Bayes" or "mle" approach. Details in \code{?abn::buildScoreCache()}.}

\item{no.cores}{integer of number of cores to parallelise.}

\item{score}{character of score that is used to score the network. Details in \code{?abn::mostprobable()}.}

\item{randomeffect}{name of grouping variable as character string.}

\item{catcov.restriction}{Passed on to \code{mclogit::mblogit} \code{CatCov} argument.}

\item{rngSEED}{integer to use as seed.}
}
\value{
list of df, dist, retain, banned, mycache.maxpar, max.par, dag.maxpar, fabn.maxpar
}
\description{
Requires output from \code{bnaiaR::findOptNoParentNodes()}.
}
\examples{
\dontrun{
abnwithmaxpar42 <-
  abnwithmaxparents(
    net.scores = optnoparents42$net.scores,
    df = df42,
    dist = dist42,
    banned = banned42,
    retain = retain42,
    filenamesuffix = "4.2"
  )
}
}
