% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adb}
\alias{adb}
\title{Aneurysm Data Base Preprocessed}
\format{
A data frame with 790 obs. of  52 variables.
}
\usage{
adb
}
\description{
Aneurysm data from Geneva (AneuQuest + AneuX) harmonized, preprocessed and transformed.
Complete data only.
For details consider the vignette
\code{vignette("data_preprocessing", package = "bnaiaR")}.
}
\details{
\itemize{
  \item Current smoker if smoked (> 300 cigarettes) and continues current smoking. If formerly smoked (>300 cigarettes) and stopped at least 6 months ago. Never smoked if < 300 cigarettes smoked ever.

  \item Never had hypertension if blood pressure is less than 120/80mmHg. Hypertension of AnyType if ...
  \itemize{
    \item ... not treated BP >140/90mmHg and the patient does not take any antihypertensive treatment.
    \item ... treated and controlled BP >140/90mmHg and the patient takes antihypertensive treatment and BP is in normal range.
    \item ... treated and not controlled BP > 140/90mmHg, the patient takes antihypertensive treatment but the BP stay higher than normal BP values.
    }

 \item Positive familial history of intracranial aneurysm if one or more first
 degree relative(s) were diagnosed with IA. Probably if the relative had a stroke
 but there is no definite diagnosis on the type of stroke.

 \item Age at time of diagnosis is grouped in steps of 5y in AgeDiag.group
 \eqn{A=[0,40), B=[40,45), C=[45,50), D=[50,55), E=[55,60), F=[60,65), G=[65,max. AgeDiag), H=unruptured}
 and in steps of 10y in AgeDiag.group.coarse
 \eqn{A=[0,40), B=[40,50), C=[50,60), D=[60,max. AgeDiag)}

 \item Size grouped in \eqn{A=< 7mm}	Maximum diameter size of aneurysm less or equal to 7mm;
 \eqn{B=7-12mm}	Maximum diameter size of aneurysm between 7 and 12mm included;
 \eqn{C=13-25mm}	Maximum diameter size of aneurysm between 13 and 25mm included;
 \eqn{D=>25mm}	Maximum diameter size of aneurysm larger or equal to 25mm;
 \eqn{E=NA}	Unknown size
}
}
\keyword{datasets}
