% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adbisgc}
\alias{adbisgc}
\title{Aneurysm Data Base of multiple study centres harmonized according to ISGC.}
\format{
data frame
\describe{
  \item{ID_1}{Internal Identifier 1}
  \item{ID_2}{Internal Identifier 2}
  \item{source}{Study source abbreviation}
  \item{gender}{Patients biological sex (male, female)}
  \item{basis_of_recruitment}{Reason for data collection}
  \item{basis_of_recruitment_CaseControl}{Reason for data collection (case, control)}
  \item{age}{Age in years of patient at time of intracranial aneurysm diagnosis: if Ruptured_IA = Yes -> "Age" refers to the patient age at time of rupture. If Ruptured_IA = No -> "Age" refers to the patient age at the time of her first diagnosed IA (oldest IA).}
  \item{positive_famillial_history}{Familiar incident known (yes, no, probably). See details.}
  \item{smoking_current_former_no}{Patient classified as current, former or non-smoker. See details.}
  \item{smoking_current_notcurrent}{Wrapper of smoking_current_former_no.}
  \item{number_of_IAs}{Total number of IAs diagnosed.}
  \item{multipleIAs}{Detection of multiple IAs, binary (yes, no). Wrapper of number_of_IAs.}
  \item{hypertension}{Known hypertension, binary (no, AnyType). See details.}
  \item{hypertension_treatment}{Known hypertension treatment. See details.}
  \item{IAlocation}{Location of the detected IA}
  \item{IAruptured}{IA rupture status (yes, no)}
  \item{IAsize_at_diag}{Measure for IA size [mm]}
  \item{IAsize_at_diag_grouped}{Wrapper of IAsize_at_diag grouped. See details.}


  \item{mRS_afterSAH}{modified ranking scale at 1y after ruptured IA}
  \item{gos_afterSAH}{Glasgow outcome scale at 1y after ruptured IA}
  \item{maxFUtime}{maximum follow-up time of patients diagnosed with unruptured IA}
  \item{termination_reason}{Follow-up termination reason of patients diagnosed with unruptured IA.	}
}
}
\usage{
adbisgc
}
\description{
Processed raw data and harmonized according \link[ExplorDataISGC]{harmonize2isgc} in \code{vignette("combineISGC", package = "ExplorDataISGC")}
and \code{vignette("data_preprocessing", package = "bnaiaR")}.
This separation of the preprocessing is on top of the anonymised raw data,
an additional layer of security to ensure privacy of patient data.
While the package \code{ExplorDataISGC} will never be published, \code{bnaiaR} is intended for open access.
}
\details{
\itemize{
 \item basis_of_recruitment
 \itemize{
   \item Retrospective Symptomatic Intracranial Aneurysm	Patient diagnosed inititaly with a symptom associated with the intracranial aneurysm / Retrospectif diagnosed  prior to november 1st 2006
   \item Symptomatic Intracranial Aneurysm	Patient diagnosed inititaly with a symptom associated with the intracranial aneurysm
   \item Retrospective Incidental Intracranial Aneurysm	Patient incidentaly diagnosed with intracranial aneurysm / Retrospectif diagnosed prior to november 1st 2006
   \item Incidental Intracranial Aneurysm	Patient incidentaly diagnosed with intracranial aneurysm
   \item Retrospective Sub Arachnoid Hemorrhage	Patient diagnosed with subarachnoid bleed as a consequence of intracranial aneurysm rupture / Retrospectif diagnosed prior to november 1st 2006
   \item Sub Arachnoid Hemorrhage	Patient diagnosed with subarachnoid bleed as a consequence of intracranial aneurysm rupture
   \item Genetically linked family member	Patient screened in the context of a positive familial history initialy without intracranial aneurysm
   \item Control	Healthy volunteers no screening for intracranial aneurysm was performed
   \item Case	Patients diagnosed with intracranial aneurysm(s) with unknown status of rupture and Genetically linked family member
 }
 \item basis_of_recruitment_CaseControl
 \itemize{
   \item case	Retrospective Symptomatic Intracranial Aneurysm, Symptomatic Intracranial Aneurysm, Retrospective Incidental , Retrospective Sub Arachnoid Hemorrhage, Incidental Intracranial AneurysmIntracranial Aneurysm, Sub Arachnoid Hemorrhage, Case, Genetically linked family member (Number of IA >=1)
   \item control	Control, Genetically linked family member (Number of IA= NA)
   \item NA	not available
 }
 \item age
 \itemize{
   \item 0 to 130	age in years
   \item NA	not available
 }
 \item positive_famillial_history
 \itemize{
   \item no	No 1st degree relative with intracranial aneurysm
   \item yes	One or more 1st degree relative(s) with intracranial aneurysm
   \item probably	The relative had a stroke but there is no definite diagnosis on the type of stroke
   \item NA	Unknown familial history of intracranial aneurysm
 }
 \item smoking_current_former_no
 \itemize{
   \item No	Never smoked (no more than 300 cigarets ever)
   \item Former	Smoked (more than 300 cigarets) and stopped (at least 6 month ago)
   \item Current	Smoked (more than 300 cigarets) and continues current smoking
   \item NA	Smoker status not available
 }
 \item smoking_current_notcurrent
 \itemize{
   \item Current	Never smoked (no more than 300 cigarets ever)
   \item Not Current	No + Former.
   \item NA	Smoker status not available
 }
 \item number_of_IAs
 \itemize{
   \item 0:inf	Number of intracranial aneurysms diagnosed
   \item NA	Not available or Healthy volunteers that were not screened are considered with an unknown number of aneurysm
 }
 \item multipleIAs
 \itemize{
   \item yes	>1 number of IAs
   \item no	0 or 1 number of IAs
   \item NA	Number of IAs not available
 }
 \item hypertension
 \itemize{
   \item no	Blood pressure less than 120/80 mm Hg
   \item yes	Blood pressure greater than 140/90 mm Hg
   \item NA	Not available
 }
 \item hypertension_treatment
 \itemize{
   \item No	No antihypertensive treatment
   \item yes	Hypertensive treatment
   \item NA	Not available
 }
 \item IAlocation. For details see \link[ExplorDataISGC]{harmonize_loc}
 \itemize{
   \item Acom	Ant communicating artery, Comm anterior CoA forward, Comm anterior CoA forward-upward, Comm anterior CoA upward, Comm anterior CoA backward, Comm anterior CoA down
   \item MCA	Sylvian bifurcation, M1 segment middle cerebral artery, M1 perforator artery,, Middle cerebral bifurcation, Middle cerebral MCA main trunk
   \item Pcom	Posterior Comm
   \item ICA	Anterior and superior wall carotid, Carotid bifurcation, Ant Choroidal segment carotid, Superior wall ICA, Lateral wall ICA CoP, Carotid bifurcation ICA bif, Lateral wall ICA ChA
   \item OphtICA	Opthalmic artery, Ophthalmic segment carotid, Medial wall carotid, Medial wall ICA ophtalmic, Medial wall ICA distal, Inferior wall ICA
   \item Basilar	Basilar Tip, Basilar tip, basilar artery bifurcation, Basilar bifurcation, P 1 posterior cerebral, P1 Posterior cerebral artery
   \item V-B	Basilar trunk, V4 segment vertebral artery, AICA, Superior cerebellar artery, PICA, Vertebral, Vertebral trunk, Basilar others, Basilar SCA, Vertebral = PICA origin, PICA distal"
   \item A2	Pericallosal cerebral artery, A2 segment ant, Pericallosal proximal, Pericallosal typical
   \item A1 segment ant	A1 segment ant
   \item PC	P2 posterior cerebral artery, P1-P2 junction posterior cerebral artery, P 2 posterior cerebral, P 1/2 posterior cerebral
   \item Other	Distal to sylvian bifurcation, Distal ant cerebral artery, Distal posterior cerebral artery, Middle cerebral peripheral, Pericallosal distal, P 3 posterior cerebral distal
   \item CavICA	Intracavernous internal carotid
   \item Infundibulum
   \item Removed	Cervical internal carotid, Petrous pyramidal
   \item NA	Unknown
 }
 \item IAruptured
 \itemize{
   \item Yes	Patient known to have ruptured an intracranial aneurysm: Ruptured, Fully treated, Symptomatic
   \item No	Patient known to have a never ruptured aneurysm: Unruptured
   \item NA	Patient with unknown status of rupture of aneurysm(s)
 }
 \item IAsize_at_diag
 \itemize{
   \item 0 to inf	Maximum diameter size in mm single digit numerical display
   \item NA	Healthy volunteers that were not screened are considered with an unknown aneurysm size
 }
 \item IAsize_at_diag_grouped
 \itemize{
   \item A=< 7mm	Maximum diameter size of aneurysm less or equal to 7mm
   \item B=7-12mm	Maximum diameter size of aneurysm between 7 and 12mm included
   \item C=13-25mm	Maximum diameter size of aneurysm between 13 and 25mm included
   \item D=>25mm	Maximum diameter size of aneurysm larger or equal to 25mm
   \item E=NA	Unknown size
 }
 \item maxFUtime
 \itemize{
   \item 0 to inf	time in years
   \item NA	Not available
 }
 \item termination_reason
 \itemize{
   \item IAex	IA excluded and no longer under observation.
   \item aSAH	IA ruptured.
   \item lossFU	Loss on follow-up.
   \item deathOC	Patient died from other cause.
   \item NA	Not available
 }
}
}
\keyword{datasets}
