% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{arc.stren.threshold}
\alias{arc.stren.threshold}
\title{Arc strength significance threshold}
\usage{
arc.stren.threshold(strength, method = "l1")
}
\arguments{
\item{strength}{Arc frequency of consensus DAG.}

\item{method}{Currently only "l1" supported.}
}
\value{
arc strength threshold
}
\description{
Compute the significance threshold for Friedman's confidence.
Deliberately copied and adapted from bnlearn source code on CRAN Github
https://github.com/cran/bnlearn/blob/0e4d3af6ad579b79bc9959e77385ec6b825ca6fc/R/arc.strength.R#L420
}
\examples{
\dontrun{dag.mcmc.boot.stren <- as.vector(round(dag.mcmc.boot, 3))
arc.stren.sign.threshold <-arc.stren.threshold(dag.mcmc.boot.stren)
# Plot relative arc strength
plot(ecdf(dag.mcmc.boot.stren))
# Draw arc strength threshold
abline(v = arc.stren.sign.threshold, lty=2)
# Draw 50\% threshold mark
abline(v=0.5)
}
}
