% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{BSF}
\alias{BSF}
\title{Balanced Scoring Function (BSF)}
\usage{
BSF(estGraph, trueGraph)
}
\arguments{
\item{estGraph}{Object of class "bn" representing the estimated network structure.}

\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}
}
\value{
integer
}
\description{
The BSF score can be used in conjunction with other traditional metrics to
provide an alternative and unbiased assessment about the capability of a
structure learning algorithm in discovering causal or BN graphs.
}
\references{
Constantinou et al. 2020 http://arxiv.org/abs/1905.12666
}
