% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{connectDBstack}
\alias{connectDBstack}
\title{Connect to DB Stack}
\usage{
connectDBstack(
  dbname,
  username,
  password,
  default.file = getOption("DB_CONFIG")
)
}
\arguments{
\item{dbname}{str with data base name}

\item{username}{str with data base user name}

\item{password}{str of password for respective data base user}

\item{default.file}{with DB configurations}
}
\value{
MariaDBConnection
}
\description{
Connect to DB Stack
}
\details{
Data base configurations must be provided as default.file in global
options as "DB_CONFIG".
}
