% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{consensusDAG}
\alias{consensusDAG}
\title{Consensus Model}
\usage{
consensusDAG(object, consensusMethod = "signEdges")
}
\arguments{
\item{object}{list of abnFit objects.}

\item{consensusMethod}{pruning strategy. "signEdge" calculates the consensus model based on accepting only significant edges.}
}
\value{
Consensus model as fitAbn object.
}
\description{
Supply as \code{object} an estimate of an initial "best" Graph $\eqn{G_0}$ from the original, real-world data $\eqn{D_0}$ to \code{paramBootAbn()}.
Let $N$ the number of bootstrap samples that are provided as \code{n.sim}.
\code{paramBootAbn()} then simulates N data set $\eqn{D_1}, \dots, \eqn{D_n}$ from model $\eqn{G_0}$
and estimates N new models $\eqn{G_1}, \dots, \eqn{G_n}$ from the respective simulated data.
}
\details{
The output from \code{paramBootAbn()} can be fed in \code{consensusDAG} which will estimate the final model $\eqn{G_{sum}}$, by summarising over all $\eqn{G_1}, \dots, \eqn{G_n}$.
The consensus model $\eqn{G_{consensus}}$ is the trimmed $\eqn{G_{sum}}$ based on accepting only significant edges.
}
\examples{
consDAGlist <- consensusDAG(object = dags,
                            consensusMethod = "signEdges")
plotAbn(consDAGlist[["consDAG"]], data.dists = dist51)

}
