% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{crim2ISGC}
\alias{crim2ISGC}
\title{Harmonization of CRIM 4.0 and 5.0 ADB scheme to ISGC}
\usage{
crim2ISGC(
  df = aneuquest.raw,
  IDcol1 = "patientID",
  IDcol2 = "aneuUniqueID",
  gendcol = "patSex",
  borcol = "recruBasis",
  agecol = "aneuReportPatAge",
  posfamhistcol = "prf_PosFamHis",
  smokingcol = "prf_SmokedTobacco",
  IAnumcol = "IAnum",
  multIAcol = NULL,
  hbpcol = "prf_HyperTension",
  hbptreatcol = "prf_HyperTension",
  IAloccol = "aneuLoca",
  IAruptcol = "ruptureStatus",
  IAsizecol = "maxDiam",
  mRScol = NULL,
  goscol = NULL,
  maxFUtimecol = NULL,
  termreascol = NULL,
  source = "geneva_aneuquest",
  returnall = FALSE
)
}
\arguments{
\item{df}{data.frame of raw data set.}

\item{IDcol1}{character string with indiviudal patient identifier or label for link with genomes of the respective study source}

\item{IDcol2}{optional other label to identify cases in original study}

\item{gendcol}{character string of column name for biological sex}

\item{borcol}{character string of column name for reason of patient's recruitment}

\item{agecol}{character string of column name for age of patient at time of intracranial aneurysm diagnosis: if Ruptured_IA = Yes -> "Age" refers to the patient age at time of rupture. If Ruptured_IA = No -> "Age" refers to the patient age at the time of her first diagnosed IA (oldest IA).}

\item{posfamhistcol}{character string of column name for positive familial history}

\item{smokingcol}{character string of column name for smoking behaviour}

\item{IAnumcol}{character string of column name for total number of diagnosed IAs per patient}

\item{multIAcol}{character string of column name for wrapper of number of IAs.}

\item{hbpcol}{character string of column name for hypertension y/n}

\item{hbptreatcol}{character string of column name for type of hypertension treatment}

\item{IAloccol}{character string of column name for IA location by vessel}

\item{IAruptcol}{character string of column name for IA rupture status}

\item{IAsizecol}{character string of column name for IA size}

\item{mRScol}{character string of column name for for modified ranking score}

\item{goscol}{character string of column name for Glasgow outcome scale}

\item{maxFUtimecol}{character string of column name for max. follow-up time}

\item{termreascol}{character string of column name for follow-up termination reason}

\item{source}{Study source abbreviation / Description}

\item{returnall}{if TRUE the original variables remain in the returned data frame.}
}
\value{
data.frame of variables in ISGC structure
}
\description{
Query the ADB for all relevant features to be harmonized into ISGC.
}
