% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{crim5toISGC}
\alias{crim5toISGC}
\title{Harmonization of CRIM 5.0 ADB scheme to ISGC}
\usage{
crim5toISGC(data, returnall = FALSE)
}
\arguments{
\item{data}{data.frame of extracted variables of ADB}

\item{returnall}{binary. If TRUE a all original variables is returned additionally to the ISGC features¨.}
}
\value{
data.frame of variables in ISGC structure
}
\description{
Query the ADB for all relevant features to be harmonized into ISGC.
}
