% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cv.metrics}
\alias{cv.metrics}
\title{Prediction Error Metrics}
\usage{
cv.metrics(xval)
}
\arguments{
\item{xval}{object returned from `bnlearn::bn.cv()`}
}
\value{
list of error metrics.
}
\description{
From cross-validation object, calculate a set of prediction error metrics.
}
\examples{
\dontrun{xval.rupture.bic <- bn.cv(
data = data,
bn = avgnet$tabu.th03,
method = "hold-out",
runs = 50,
fit = "mle",
loss = "pred",
loss.args = list(target = "Ruptured_IA")
)
cv.metrics(xval.rupture.bic)}
}
