% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{dagswithdiffparentsplt}
\alias{dagswithdiffparentsplt}
\title{Draw graphs for different number of max parents.}
\usage{
dagswithdiffparentsplt(
  net.scores.dags,
  df,
  SAVEPLOTS,
  filenamesuffix,
  filenamebase = FILENAMEbase,
  filename = FILENAME,
  filetype = "pdf",
  plotwidth = PLOTWIDTH,
  plotheight = PLOTHEIGHT
)
}
\arguments{
\item{net.scores.dags}{output from \code{bnaiaR::findOptNoParentNodes()$net.scores.dags}}

\item{df}{data.frame with training data}

\item{SAVEPLOTS}{If TRUE the plot will be saved.}

\item{filenamesuffix}{character specifying the current run (e.g. experiment number).}

\item{filenamebase}{character of path to location where results should be stored.}

\item{filename}{character specifying the current set of runs.}

\item{filetype}{"pdf" or "png" or any that is supported by \code{?ggplot2::ggsave()}.}

\item{plotwidth}{integer specifying plot width.}

\item{plotheight}{integer specifying plot height.}
}
\value{
plot
}
\description{
Draw graphs for different number of max parents.
}
