% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{DDM.constantinou}
\alias{DDM.constantinou}
\title{DAG Dissimilarity Metric (DDM)}
\usage{
DDM.constantinou(estGraph, trueGraph)
}
\arguments{
\item{estGraph}{Object of class "bn" representing the estimated network structure.}

\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}
}
\value{
integer
}
\description{
DAG Dissimilarity Metric (DDM)
}
\references{
Constantinou et al. 2020 http://arxiv.org/abs/1905.12666
}
