% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_bnaiar_compl}
\alias{df_bnaiar_compl}
\title{BnaiaR variables without NAs in different transformations}
\format{
A data.frame
}
\usage{
df_bnaiar_compl
}
\description{
This is the basis on which the experiment-data was prepared.
}
\details{
For details consider the vignette
\code{vignette("data_preparation_for_experiments", package = "bnaiaR")}.
}
\keyword{datasets}
