% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{discrete_bns}
\alias{discrete_bns}
\title{Discrete Bayesian Network Models}
\format{
A list of 2 lists with different "bn" objects and "bn.strength"
objects respectively.
}
\usage{
discrete_bns
}
\description{
For details consider the vignette
\code{vignette("discrete_BN_SL", package = "bnaiaR")}.
}
\keyword{datasets}
