% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ekg}
\alias{ekg}
\title{Evidence Knowledge Graph}
\format{
A list of 3 items and class "bn".
}
\usage{
ekg
}
\description{
For details consider the vignette
\code{vignette("evidenceknowledgegraph", package = "bnaiaR")}.
}
\keyword{datasets}
