% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp11_dat}
\alias{exp11_dat}
\title{Discrete BN Analysis Input Data}
\format{
A list of 4 items:
\describe{
  \item{abndata}{data frame of	790 obs. of  9 variables, discrete and continuous.}
  \item{dist}{List of 9. Type of distribution for each variable in abndata.}
  \item{banned}{matrix of 9x9. Prior knowledge implementation as matrix of banned arcs for package `abn`.}
  \item{bl}{data frame. Prior knowledge implementation as blacklist for package `bnlearn`.}
}
}
\usage{
exp11_dat
}
\description{
Aneurysm data from Geneva harmonized, preprocessed and transformed.
A selection of variables from \code{data("adb", package = "bnaiaR")},
prior knowledge and additional required data-set depended information for
BN analysis with `bnlearn` and `abn`.
For details consider the vignette
\code{vignette("data_preparation_for_experiments", package = "bnaiaR")}.
}
\keyword{datasets}
