% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp44_dat}
\alias{exp44_dat}
\title{Additive BN Analysis Input Data, whitelisted study_source}
\format{
A list of 4 items:
\describe{
  \item{abndata}{data frame of	7481 obs. of  10 discrete and continuous variables.}
  \item{dist}{List of 10. Type of distribution for each variable in abndata.}
  \item{banned}{matrix of 10x10. Prior knowledge implementation as matrix of banned arcs for package `abn`.}
  \item{bl}{data frame. Prior knowledge implementation as blacklist for package `bnlearn`.}
  \item{retain}{matrix of 10x10. Prior knowledge implementation as matrix of retained arcs for package `abn`.}
  \item{wl}{data frame. Prior knowledge implementation as whitelist for package `bnlearn`.}
}
}
\usage{
exp44_dat
}
\description{
Variation of \code{data(exp4_dat, package = "bnaiaR")} with `study_source`
forced (whitelisted) as parent of all other variables.
For details consider the vignette
\code{vignette("data_preparation_for_experiments", package = "bnaiaR")}.
}
\keyword{datasets}
