% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp4_dat}
\alias{exp4_dat}
\title{Additive BN Analysis Input Data}
\format{
A list of 4 items:
\describe{
  \item{abndata}{data frame of	7481 obs. of  10 discrete and continuous variables.}
  \item{dist}{List of 10. Type of distribution for each variable in abndata.}
  \item{banned}{matrix of 10x10. Prior knowledge implementation as matrix of banned arcs for package `abn`.}
  \item{bl}{data frame. Prior knowledge implementation as blacklist for package `bnlearn`.}
  \item{retain}{matrix of 10x10. Prior knowledge implementation as matrix of retained arcs for package `abn`.}
  \item{wl}{data frame. Prior knowledge implementation as whitelist for package `bnlearn`.}
}
}
\usage{
exp4_dat
}
\description{
Aneurysm data from multiple centres harmonized, preprocessed and transformed.
A selection of variables from \code{data("adbisgc", package = "bnaiaR")},
prior knowledge and additional required data-set depended information for
BN analysis with `bnlearn` and `abn`.
For details consider the vignette
\code{vignette("data_preparation_for_experiments", package = "bnaiaR")}.
}
\keyword{datasets}
