% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{f1}
\alias{f1}
\title{F1 Score of Graphs}
\usage{
f1(estGraph, trueGraph)
}
\arguments{
\item{estGraph}{Object of class "bn" representing the estimated network structure.}

\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}
}
\value{
integer
}
\description{
F1 Score of Graphs
}
\references{
Constantinou et al. 2020 http://arxiv.org/abs/1905.12666
}
