% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{findOptNoParentNodes}
\alias{findOptNoParentNodes}
\title{Find Optimal Number of Parent Nodes}
\usage{
findOptNoParentNodes(
  df,
  dist,
  banned,
  retain,
  filenamesuffix,
  filenamebase = FILENAMEbase,
  filename = FILENAME,
  method = METHOD,
  no.cores = n.cores,
  randomeffect = NULL,
  catcov.restriction = "diagonal",
  score = SCORE,
  rngSEED = SEED
)
}
\arguments{
\item{df}{data.frame with training data}

\item{dist}{list of variable = distribution type. Details in \code{?abn::buildScoreCache()}.}

\item{banned}{matrix with column variable as parent of row-variable, that defines which arcs are not permitted. Details in \code{?abn::buildScoreCache()}.}

\item{retain}{matrix with column variable as parent of row-variable, that defines which arcs are enforced. Details in \code{?abn::buildScoreCache()}.}

\item{filenamesuffix}{character specifying the current run (e.g. experiment number).}

\item{filenamebase}{character of path to location where results should be stored.}

\item{filename}{character specifying the current set of runs.}

\item{method}{either "Bayes" or "mle" approach. Details in \code{?abn::buildScoreCache()}.}

\item{no.cores}{integer of number of cores to parallelise.}

\item{randomeffect}{name of grouping variable as character string.}

\item{catcov.restriction}{Passed on to \code{mclogit::mblogit} \code{CatCov} argument.}

\item{score}{character of score that is used to score the network. Details in \code{?abn::mostprobable()}.}

\item{rngSEED}{integer to use as seed.}
}
\value{
list of df, dist, retain, banned, net.score, net.score.dags.
}
\description{
Find Optimal Number of Parent Nodes
}
\examples{
\dontrun{
optnoparents42 <-
findOptNoParentNodes(
  df = df42,
  dist = dist42,
  banned = banned42,
  retain = retain42,
  filenamesuffix = "4.2"
)
}
}
