% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customSLscore.R
\name{glmm.bic}
\alias{glmm.bic}
\title{Additive discrete Bayesian network mixed-effects model}
\usage{
glmm.bic(node, parents, data, args)
}
\arguments{
\item{node}{character string of current node. Corresponds to variable name in data.}

\item{parents}{character vector of the node's parents.}

\item{data}{data frame for SL incl. the mixed effect variable.}

\item{args}{list containing node distribution (dist) and variable name for mixed effect (mixed_effect)}
}
\value{
model's score (BIC) as integer
}
\description{
Custom score function to regress nodes on their parents as generalized linear mixed model.
}
\examples{
\dontrun{
model <- glmm.bic(node = "IAlocation_group",
                  parents = c("gender", "hpt_aware", "IAsize_diag_grouped_merged","multipleIAs"),
                  data = data.dbn.additive,
                  args = list(dist = dist.dbn.additive,
                              mixed_effect = "study_source"))

# or for bootstrapping
bb = boot.strength(data.dbn.additive,
algorithm = "tabu",
R = 500,
cluster = cl,
algorithm.args = list(score = "custom",
                      fun = glmm.bic,
                      blacklist = bl.dbn.additive,
                      # whitelist = wl.dbn.allstudy,
                      args = list(dist = dist.dbn.additive,
                                  mixed_effect = "study_source"),
                      tabu = 50))
}
}
\references{
Scutari, Marco, Christopher Marquis, and Laura Azzimonti. “Using Mixed-Effects Models to Learn Bayesian Networks from Related Data Sets.” In Proceedings of The 11th International Conference on Probabilistic Graphical Models, 73–84. PMLR, 2022. https://proceedings.mlr.press/v186/scutari22a.html.

Azzimonti, Laura, Giorgio Corani, and Marco Scutari. “A Bayesian Hierarchical Score for Structure Learning from Related Data Sets.” International Journal of Approximate Reasoning 142 (March 1, 2022): 248–65. https://doi.org/10.1016/j.ijar.2021.11.013.
}
