% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmonize_loc}
\alias{harmonize_loc}
\title{Harmonize IA locations by vessel to a unified standard}
\usage{
harmonize_loc(IAloc)
}
\arguments{
\item{IAloc}{data frame column of IA locations by vessels from study raw data}
}
\value{
data frame column of IA locations by vessels ISGC
}
\description{
The unified standard was selected by a least common denominator over all data sources.
For details see: (DataDescription_ISGC_Phenotypes_BayesianModelling)[https://zhaw.sharepoint.com/:x:/s/Aneurysm/EeALEWUBLvRKnnNxZo2PO4MBF37FnRa_7BrJu_aRiLtVFw?e=AIlewJ]
}
\examples{
\dontrun{
dfnantes <- read_xlsx(system.file("extdata/nantes/ICAN clinical database for ISGC.xlsx", package = "ExplorDataISGC"))
# Create new features
dfnantes_raw <- dfnantes \%>\%
  # Create ID2 as unique patient identifier (see email from Romain Bourcier on 20.5.22)
    mutate(ID2 = 1:n()) \%>\%
  # Create age at diagnosis in years
    mutate(ageatdiagy =trunc(lubridate::time_length(
      lubridate::interval(
        start = lubridate::as_date(age),
        end = lubridate::as_date(`age at time of intracranial aneurysm (IA) diagnosis or at time of subarachnoid hemorrhage (SAH)`)),
        unit = "year"))) \%>\%
  # translate locations from numbers to names
    mutate(across(contains("Location"),
                  .fns = nanteslocnumber2locname,
                  .names = "vessel_{.col}")) \%>\%
  # harmonize locations to isgc groups
    mutate(across(contains("vessel_"),
                  .fns = harmonize_loc,
                  .names = "isgc_{.col}")) \%>\%
  # group locations by risk
    mutate(across(contains("isgc_vessel_"),
                  .fns = harmonize_locbyrisk,
                  .names = "locrisk_{.col}"))
}
}
