% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmonize_locbyrisk}
\alias{harmonize_locbyrisk}
\title{Group IA (ISGC) locations by their risk of rupture}
\usage{
harmonize_locbyrisk(IAlocation)
}
\arguments{
\item{IAlocation}{data frame column of IA locations as ISGC-vessels}
}
\value{
data frame column of IA locations by risk of rupture (3 = high risk)
}
\description{
Group IA (ISGC) locations by their risk of rupture
}
\examples{
\dontrun{
dfnantes <- read_xlsx(system.file("extdata/nantes/ICAN clinical database for ISGC.xlsx", package = "ExplorDataISGC"))
# Create new features
dfnantes_raw <- dfnantes \%>\%
  # Create ID2 as unique patient identifier (see email from Romain Bourcier on 20.5.22)
    mutate(ID2 = 1:n()) \%>\%
  # Create age at diagnosis in years
    mutate(ageatdiagy =trunc(lubridate::time_length(
      lubridate::interval(
        start = lubridate::as_date(age),
        end = lubridate::as_date(`age at time of intracranial aneurysm (IA) diagnosis or at time of subarachnoid hemorrhage (SAH)`)),
        unit = "year"))) \%>\%
  # translate locations from numbers to names
    mutate(across(contains("Location"),
                  .fns = nanteslocnumber2locname,
                  .names = "vessel_{.col}")) \%>\%
  # harmonize locations to isgc groups
    mutate(across(contains("vessel_"),
                  .fns = harmonize_loc,
                  .names = "isgc_{.col}")) \%>\%
  # group locations by risk
    mutate(across(contains("isgc_vessel_"),
                  .fns = harmonize_locbyrisk,
                  .names = "locrisk_{.col}"))
}
}
