% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmonize_smokingCFN}
\alias{harmonize_smokingCFN}
\title{Harmonize Smoking (three levels) values}
\usage{
harmonize_smokingCFN(smoking)
}
\arguments{
\item{smoking}{as character string with three possible levels: current, former, none smoker}
}
\value{
character string
}
\description{
See also \code{\link[=harmonize_smokingCN]{harmonize_smokingCN()}}.
}
