% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{harmonize_smokingCN}
\alias{harmonize_smokingCN}
\title{Harmonize Smoking (two levels) values}
\usage{
harmonize_smokingCN(smoking)
}
\arguments{
\item{smoking}{as character string with two possible levels: current, none smoker}
}
\value{
character string
}
\description{
See also \code{\link[=harmonize_smokingCFN]{harmonize_smokingCFN()}}.
}
