% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{maxparentsplot}
\alias{maxparentsplot}
\title{Maximum Number of Parents for ABN}
\usage{
maxparentsplot(
  net.scores,
  SAVEPLOTS,
  filenamesuffix,
  filenamebase = FILENAMEbase,
  filename = FILENAME,
  filetype = "pdf",
  plotwidth = PLOTWIDTH,
  plotheight = PLOTHEIGHT
)
}
\arguments{
\item{net.scores}{output from \code{bnaiaR::findOptNoParentNodes()$net.scores}}

\item{SAVEPLOTS}{If TRUE the plot will be saved.}

\item{filenamesuffix}{character specifying the current run (e.g. experiment number).}

\item{filenamebase}{character of path to location where results should be stored.}

\item{filename}{character specifying the current set of runs.}

\item{filetype}{"pdf" or "png" or any that is supported by \code{?ggplot2::ggsave()}.}
}
\value{
ggplot2 and data.frame.
}
\description{
Evaluate the number of parent nodes that does not increase the score anymore.
}
