% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{mcmcabn_bnaiar}
\alias{mcmcabn_bnaiar}
\title{Run MCMC for ABN}
\usage{
mcmcabn_bnaiar(
  mycache.maxpar,
  max.par,
  dag.maxpar,
  fabn.maxpar,
  df,
  dist,
  banned,
  retain,
  filenamesuffix,
  filenamebase = FILENAMEbase,
  filename = FILENAME,
  method = METHOD,
  no.cores = n.cores,
  CLTYPE = "FORK",
  score = SCORE,
  mcmcseeds = MCMC.SEEDS,
  mcmcscheme = MCMC.SCHEME,
  mcmcprior = MCMC.PRIOR,
  prob.rev = PROB.REV,
  prob.mbr = PROB.MBR
)
}
\arguments{
\item{mycache.maxpar}{output from \code{bnaiaR::abnwithmaxparents()$mycache.maxpar}.}

\item{max.par}{output from \code{bnaiaR::abnwithmaxparents()$maxpar}.}

\item{dag.maxpar}{output from \code{bnaiaR::abnwithmaxparents()$dag.maxpar}.}

\item{fabn.maxpar}{output from \code{bnaiaR::abnwithmaxparents()$fabn.maxpar}. Only used to collect in the final results.}

\item{df}{data.frame with training data}

\item{dist}{list of variable = distribution type. Details in \code{?abn::buildScoreCache()}.}

\item{banned}{matrix with column variable as parent of row-variable, that defines which arcs are not permitted. Details in \code{?abn::buildScoreCache()}.}

\item{retain}{matrix with column variable as parent of row-variable, that defines which arcs are enforced. Details in \code{?abn::buildScoreCache()}.}

\item{filenamesuffix}{character specifying the current run (e.g. experiment number).}

\item{filenamebase}{character of path to location where results should be stored.}

\item{filename}{character specifying the current set of runs.}

\item{method}{either "Bayes" or "mle" approach. Details in \code{?abn::buildScoreCache()}.}

\item{no.cores}{integer of number of cores to parallelise.}

\item{CLTYPE}{"FORK" (default) or "PSOCK" if on windows. The later can cause memory issues.}

\item{score}{character of score that is used to score the network. Details in \code{?abn::mostprobable()}.}

\item{mcmcseeds}{vector of integers who set the seeds in each parallel run. Details in \code{?mcmcabn::mcmcabn()}.}

\item{mcmcscheme}{vector of integers with the number of returned DAGs, the number of thinned steps and length of the burn-in phase. Details in \code{?mcmcabn::mcmcabn()}.}

\item{mcmcprior}{integer of 1 for uniform structural prior and 2 for weighted prior. Details in \code{?mcmcabn::mcmcabn()}.}

\item{prob.rev}{probability of selecting a new edge reversal. Details in \code{?mcmcabn::mcmcabn()}.}

\item{prob.mbr}{probability of selecting a Markov blanket resampling scheme. Details in \code{?mcmcabn::mcmcabn()}.}
}
\value{
list of df, dist, retain, banned, score, method, mycache.maxpar, max.par, dag.maxpar, fabn.maxpar, mcmc.out.list, mcmc.scheme, prob.rev, prob.mbr, prior.choice.
}
\description{
Requires output from \code{bnaiaR::abnwithmaxparents()}.
}
\examples{
\dontrun{
mcmcabn42 <-
mcmcabn_bnaiar(
  mycache.maxpar = abnwithmaxpar42$mycache.maxpar,
  max.par = abnwithmaxpar42$max.par,
  dag.maxpar = abnwithmaxpar42$dag.maxpar,
  fabn.maxpar = abnwithmaxpar42$fabn.maxpar,
  df = df42,
  dist = dist42,
  banned = banned42,
  retain = retain42,
  filenamesuffix = "4.2"
)
}
}
