% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{nanteslocnumber2locname}
\alias{nanteslocnumber2locname}
\title{Dictionary translating IA location numbers to vessel names}
\usage{
nanteslocnumber2locname(loc)
}
\arguments{
\item{loc}{data frame column of IA locations in numbers}
}
\value{
data frame column of IA locations as vessel names
}
\description{
See email from Romain Bourcier on 20.5.22
}
\examples{
\dontrun{
dfnantes <- read_xlsx(system.file("extdata/nantes/ICAN clinical database for ISGC.xlsx", package = "ExplorDataISGC"))
# Create new features
dfnantes_raw <- dfnantes \%>\%
  # Create ID2 as unique patient identifier (see email from Romain Bourcier on 20.5.22)
    mutate(ID2 = 1:n()) \%>\%
  # Create age at diagnosis in years
    mutate(ageatdiagy =trunc(lubridate::time_length(
      lubridate::interval(
        start = lubridate::as_date(age),
        end = lubridate::as_date(`age at time of intracranial aneurysm (IA) diagnosis or at time of subarachnoid hemorrhage (SAH)`)),
        unit = "year"))) \%>\%
  # translate locations from numbers to names
    mutate(across(contains("Location"),
                  .fns = nanteslocnumber2locname,
                  .names = "vessel_{.col}")) \%>\%
  # harmonize locations to isgc groups
    mutate(across(contains("vessel_"),
                  .fns = harmonize_loc,
                  .names = "isgc_{.col}")) \%>\%
  # group locations by risk
    mutate(across(contains("isgc_vessel_"),
                  .fns = harmonize_locbyrisk,
                  .names = "locrisk_{.col}"))
}
}
