% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{network.metrics}
\alias{network.metrics}
\title{Calculate Network metrics during parameter analysis}
\usage{
network.metrics(
  data,
  estGraph,
  estGraph.th0,
  algo,
  tabulistsize = NULL,
  b = NULL,
  trueGraph = NULL,
  crt = NULL,
  priorname = NULL
)
}
\arguments{
\item{data}{DataFrame supplied to the structure learning algorithm.}

\item{estGraph}{Object of class "bn" representing the estimated network structure on the estimated confidence level.}

\item{estGraph.th0}{Object of class "bn" representing the estimated network structure on the estimated confidence level set to zero.}

\item{algo}{character string of the structure learning algorithm.}

\item{tabulistsize}{integer if tabu algorithm was used, tabu list size.}

\item{b}{integer. Number of bootstrap replicates}

\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}

\item{crt}{character string of the criterion/score used by the structure learning algorithm.}

\item{priorname}{charachter string of the prior.}
}
\value{
DataFrame of various network metrics.
}
\description{
See details bewlow for a list of the currently implemented network metrics.
}
\details{
Currently implemented network metrics.
\itemize{
 \item{Score/Criterion}{Score/Criterion used by SL Algorithm}
 \item{Netscore}{Value of Network score}
 \item{prior}{Prior applied to network score}
 \item{threshold}{Arc strength/confidence threshold}
 \item{Number of arcs}{Number of detected arcs including arcs below threshold}
 \item{Significant arcs}{Number of significant arcs.}
 }
If a reference graph is supplied the following metrics are calculated:
\itemize{
 \item{True negatives}{}
 \item{True positives}{}
 \item{False negatives}{}
 \item{False positives}{}
 \item{False positive rate}{}
 \item{True positive rate}{}
 \item{Accuracy}{}
 \item{Structural Hamming Distance (SHD)}{}
 \item{Balanced Scoring Function}{}
 \item{Bayesian factor}
}
}
