% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorplots.R
\name{overview_plt}
\alias{overview_plt}
\title{Single Variabel Overview Plot}
\usage{
overview_plt(df, x, decreasing = FALSE, warnings = FALSE)
}
\arguments{
\item{df}{data.frame()}

\item{x}{str. Columnname.}

\item{decreasing}{is set to FALSE, columns are not sorted. If TRUE they are arranged in
decreasing order.}

\item{warnings}{prints warnings if TRUE, FALSE suppresses warnings.}
}
\value{
ggplot object
}
\description{
Single Variabel Overview Plot
}
