% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorplots.R
\name{overview_plt_max}
\alias{overview_plt_max}
\title{Multivariable Overview Plot}
\usage{
overview_plt_max(df)
}
\arguments{
\item{df}{data.frame()}
}
\value{
ggplot object
}
\description{
Multivariable Overview Plot
}
