% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorplots.R
\name{overview_plt_max_grouped}
\alias{overview_plt_max_grouped}
\title{Multivariable Overview Plot colored by source}
\usage{
overview_plt_max_grouped(
  df,
  colorby = "study_source",
  ignore.idx = T,
  cont.vars = c("age_diag", "IAsize_diag", "IAsize_diag_log"),
  disc.cont.ratio = c(4, 1)
)
}
\arguments{
\item{df}{data frame}

\item{colorby}{char. str. of variable name to which the color should be applied}

\item{ignore.idx}{if TRUE the index variable will be ignored}

\item{cont.vars}{str. of continuous variable's names.}

\item{disc.cont.ratio}{relative number of rows of discrete and continuous plots.}
}
\value{
ggplot2
}
\description{
Multivariable Overview Plot colored by source
}
