% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedadditiveBNs.R
\name{paramBootAbn}
\alias{paramBootAbn}
\title{Parameteric Bootstrapping}
\usage{
paramBootAbn(
  object,
  n.sim,
  dag.banned,
  dag.retained,
  max.parents,
  btseeds,
  verbose = FALSE
)
}
\arguments{
\item{object}{of class abnFit}

\item{n.sim}{number of bootstrap iterations.}

\item{dag.banned}{a matrix or a formula statement defining which arcs are not permitted - banned. Note that colnames and rownames must be set, otherwise same row/column names as data.df will be assumed. If set as NULL an empty matrix is assumed.}

\item{dag.retained}{a matrix or a formula statement defining which arcs are must be retained in any model search. Note that colnames and rownames must be set, otherwise same row/column names as data.df will be assumed. If set as NULL an empty matrix is assumed.}

\item{max.parents}{a constant or named list giving the maximum number of parents allowed.}

\item{btseeds}{integer vector with individual seeds for each simulation. Requires one, unique seed for each iteration.}

\item{verbose}{print more output.}
}
\value{
list of abn fit objects
}
\description{
Supply as \code{object} an estimate of an initial "best" Graph $\eqn{G_0}$ from the original, real-world data $\eqn{D_0}$.
Let $N$ the number of bootstrap samples that are provided as \eqn{n.sim}.
This function then simulates N data set $\eqn{D_1}, \dots, \eqn{D_n}$ from model $\eqn{G_0}$
and estimates N new models $\eqn{G_1}, \dots, \eqn{G_n}$ from the respective simulated data.
}
\details{
The output can be fed in \code{consensusDAG} which will estimate the final model $\eqn{G_{sum}}$, by summarising over all $\eqn{G_1}, \dots, \eqn{G_n}$.
The consensus model $\eqn{G_{consensus}}$ is the trimmed $\eqn{G_{sum}}$ based on accepting only significant edges.
}
\examples{
dags <- paramBootAbn(object = myres,
                     n.sim = 100,
                     dag.banned = banned51[-10, -10], # all except study_source
                     dag.retained = retain51[-10, -10], # all except study_source
                     max.parents = 4,
                     btseeds = c(1:100)*SEED)

}
