% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotROCAUC}
\alias{plotROCAUC}
\title{Plot and save ROC curve with AUC}
\usage{
plotROCAUC(roc, aucCI, FILENAME, PLOTPATH = NULL, SAVE = SAVEPLOTS)
}
\arguments{
\item{roc}{object returned from `pROC::roc`}

\item{aucCI}{object returned from `pROC::ci.auc`}

\item{FILENAME}{string of filename}

\item{PLOTPATH}{string of path}

\item{SAVE}{If True, the plot will be saved.}
}
\value{
plot or none.
}
\description{
Plots the ROC Curve with AUC and AUC CI values.
}
\examples{
\dontrun{roc <- pROC::roc(ValidSet$Ruptured_IA, pred_glm)
auc <- roc$auc
aucCI <- round(pROC::ci.auc(ValidSet$Ruptured_IA, pred_glm), 2)
plotROCAUC(roc, aucCI, FILENAME = "logreg_IArupture_roc.png")
}
}
