% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{postBURNin}
\alias{postBURNin}
\title{Burn-in Phase after MCMCABN}
\usage{
postBURNin(mcmc.out.list, burnin.length)
}
\arguments{
\item{mcmc.out.list}{output of mcmcabn()}

\item{burnin.length}{int. of number of steps to remove.}
}
\value{
Burned mcmcabn() output.
}
\description{
Speed up computation time of mcmcabn() and gain flexibility by removing the
first n steps after the MCMC run.
}
