% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{postTHINN}
\alias{postTHINN}
\title{Thinn MCMC sample after MCMCABN}
\usage{
postTHINN(mcmc.out.list, thinningsteps)
}
\arguments{
\item{mcmc.out.list}{output of mcmcabn()}

\item{thinningsteps}{int. Thinning factor.}
}
\value{
Thinned mcmcabn() output.
}
\description{
Speed up computation time of mcmcabn() and gain flexibility by thinning
after the MCMC run.
Thinning keeps all kth step.
}
