% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep.data2plot}
\alias{prep.data2plot}
\title{Summarise data processing steps required to plot structure learning metric}
\usage{
prep.data2plot(eval, newnames = NULL)
}
\arguments{
\item{eval}{DataFrame output of network.metrics}

\item{newnames}{vector of character strings with names to display in plot.
The order must match to the order of names(eval)!}
}
\value{
tibble which can be further processed to plot
}
\description{
Summarise data processing steps required to plot structure learning metric
}
