% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{prep_exp_data}
\alias{prep_exp_data}
\title{Prepare Experiment specific data sets}
\usage{
prep_exp_data(
  dat = adb,
  EXPNO,
  age,
  location,
  size,
  smoking = "binCFN",
  SAVE = T,
  FILENAMEbase = NULL
)
}
\arguments{
\item{dat}{ADB preprocessed}

\item{EXPNO}{integer for experiment number.}

\item{age}{Type of variable: "cont", "disc.grouped-multinomial", "disc.grouped-binomial"}

\item{location}{Type of variable: "byVessel-multinomial", "byVessel-binomial", "byRisk-multinomial", "byRisk-binomial"}

\item{size}{Type of variable: "grouped-multinomial", "grouped.merged-multinomial", "log", "grouped-binomial","grouped.merged-binomial"}

\item{smoking}{Type of variable: "mult", "binCFN", "binCnC"}

\item{SAVE}{if TRUE, it the data sets were saved in FILENAMEbase}

\item{FILENAMEbase}{Place to store the files.}
}
\value{
Three data sets: Experiment data, banned matrix, list of distributions, Blacklist
}
\description{
Prepare Experiment specific data sets
}
\examples{
\dontrun{prep_exp_data(dat = adb,
EXPNO = 06,
age = "cont",
location = "byRisk-binomial",
size = "log",
SAVE = T,
FILENAMEbase = "./data/")
}
}
