% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorplots.R
\name{printorsave_dfsummary}
\alias{printorsave_dfsummary}
\title{Save or print summarytools::dfSummary()}
\usage{
printorsave_dfsummary(df, FILENAME, SAVE = SAVE, PLOTPATH = PLOTPATH)
}
\arguments{
\item{df}{data.frame.}

\item{FILENAME}{str. with file name suffix: "dfsummary_<FILENAME>"}

\item{SAVE}{TRUE if output should be stored as html file or printed to the console.}

\item{PLOTPATH}{str. with path to save. i.e. `/home/user/Documents`}
}
\value{
Data Frame Summary or NULL
}
\description{
Prints summary of `df` or saves a html document named "dfsummary_<FILENAME>" at
PLOTPATH.
}
\examples{
\dontrun{
printorsave_dfsummary(df=dfucl, FILENAME = "dfucl", SAVE = FALSE, PLOTPATH = PLOTPATH)
}
}
