% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{score.ratio}
\alias{score.ratio}
\title{Network Score Ratio}
\usage{
score.ratio(trueGraph, estGraph, data, crt)
}
\arguments{
\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}

\item{estGraph}{Object of class "bn" representing the estimated network structure.}

\item{data}{DataFrame supplied to the structure learning algorithm.}

\item{crt}{character string of the criterion/score used by the structure learning algorithm.}
}
\value{
integer
}
\description{
Structure learning maximizes the score function.
The quality of the learned network is its score.
To compare resulting networks from structure learning, on their network size
and average degree (measure of sparseness), Perrier et al. 2008 suggested to
use the score ratio of the "Optimal-Structure" over the "Learned-Structure".
This allows to compare network score across different parameters, as score
values heavily vary upon changing parameters.
}
\details{
The closer to 1, the better.
}
\references{
Perrier, Eric, Seiya Imoto, and Satoru Miyano.
“Finding Optimal Bayesian Network Given a Super-Structure.”
Journal of Machine Learning Research 9, no. 74 (2008): 2251–86.
http://jmlr.org/papers/v9/perrier08a.html.
}
