% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureanalysis.R
\name{SHD.constantinou}
\alias{SHD.constantinou}
\title{Structural Hamming Distance (Constantinou et al.)}
\usage{
SHD.constantinou(estGraph, trueGraph)
}
\arguments{
\item{estGraph}{Object of class "bn" representing the estimated network structure.}

\item{trueGraph}{Object of class "bn" representing the domain knowledge as graph ("intuition graph" or "ground truth graph").}
}
\value{
integer
}
\description{
EXPERIMENTAL!
Doesn't account for WL/BL.
}
\references{
Constantinou et al. 2020 http://arxiv.org/abs/1905.12666
}
